+++ shipgen/DPPythia8Generator.cxx
@@ -5,13 +5,14 @@
 #include "DPPythia8Generator.h"
 
 #include <cmath>
-
+#include "BeamSmearingUtils.h"
 #include "FairPrimaryGenerator.h"
 #include "Pythia8/Pythia.h"
 #include "TDatabasePDG.h"  // for TDatabasePDG
 #include "TMath.h"
 #include "TROOT.h"
 #include "TSystem.h"
+
 const Double_t cm = 10.;                  // pythia units are mm
 const Double_t c_light = 2.99792458e+10;  // speed of light in cm/sec (c_light
                                           // = 2.99792458e+8 * m/s)
@@ -28,6 +29,8 @@ DPPythia8Generator::DPPythia8Generator() {
   fFDs = 7.7 / 10.4;  // correction for Pythia6 to match measured Ds production
   fpbrem = kFALSE;
   fpbremPDF = 0;
+  fsmearBeam = 0.8;  // default value for smearing beam (8 mm)
+  fPaintBeam = 5;  // default value for painting beam (5 cm)
   fdy = kFALSE;
   fDPminM = 0.5;
   fInputFile = nullptr;
@@ -38,7 +41,7 @@ DPPythia8Generator::DPPythia8Generator() {
   // fPythiaHadDecay =  new Pythia8::Pythia();
 }
 // -------------------------------------------------------------------------
-
+void DPPythia8Generator::Print() { fPythia->settings.listAll(); };
 // -----   Default constructor   -------------------------------------------
 Bool_t DPPythia8Generator::Init() {
   if (fUseRandom1) fRandomEngine = std::make_shared<PyTr1Rng>();
@@ -46,43 +49,9 @@ Bool_t DPPythia8Generator::Init() {
   fPythia->setRndmEnginePtr(fRandomEngine);
   // fPythiaHadDecay->setRndmEnginePtr(fRandomEngine);
   fn = 0;
-  if (fextFile) {
-    /*if (0 == strncmp("/eos",fextFile,4) ) {
-     if (0 == strncmp("/eos",fextFile,4) ) {
-     TString tmp = gSystem->Getenv("EOSSHIP");
-     tmp+=fextFile;
-     fInputFile  = TFile::Open(tmp);
-     fLogger->Info(MESSAGE_ORIGIN,""Open external file with charm or beauty
-    hadrons on eos: %s",tmp->Data()); if (!fInputFile) {
-      fLogger->Fatal(MESSAGE_ORIGIN, "Error opening input file. You may have
-    forgotten to provide a krb5 token. Try kinit username@lxplus.cern.ch");
-      return kFALSE; }
-    }else{
-      fLogger->Info(MESSAGE_ORIGIN,"Open external file with charm or beauty
-    hadrons: %s",fextFile); fInputFile  = new TFile(fextFile->c_str()); if
-    (!fInputFile) { fLogger->Fatal(MESSAGE_ORIGIN, "Error opening input file");
-    return kFALSE;
-    }
-    }
-    if (fInputFile->IsZombie()) {
-     fLogger->Fatal(MESSAGE_ORIGIN, "File is corrupted");
-     return kFALSE; }
-     fTree = (TTree *)fInputFile->Get("pythia6");
-     fNevents = fTree->GetEntries();
-     fn = firstEvent;
-     fTree->SetBranchAddress("id",&hid);                // particle id
-     fTree->SetBranchAddress("px",&hpx);   // momentum
-     fTree->SetBranchAddress("py",&hpy);
-     fTree->SetBranchAddress("pz",&hpz);
-     fTree->SetBranchAddress("E",&hE);
-     fTree->SetBranchAddress("M",&hM);
-     fTree->SetBranchAddress("mid",&mid);   // mother
-     fTree->SetBranchAddress("mpx",&mpx);   // momentum
-     fTree->SetBranchAddress("mpy",&mpy);
-     fTree->SetBranchAddress("mpz",&mpz);
-     fTree->SetBranchAddress("mE",&mE);
-    */
-  } else if (!fpbrem) {
+
+
+  if (!fpbrem) {
     if (debug) {
       std::cout << "Beam Momentum " << fMom << std::endl;
     }
@@ -98,31 +67,15 @@ Bool_t DPPythia8Generator::Init() {
 
   } else {
     if (!fpbremPDF) {
-      // std::cout << " Failed in retrieving dark photon PDF for production by
-      // proton bremstrahlung! Exiting..." << std::endl;
-      LOG(fatal) << "Failed in retrieving dark photon PDF for production by "
+         LOG(fatal) << "Failed in retrieving dark photon PDF for production by "
                     "proton bremstrahlung!";
       return kFALSE;
     }
   }
-  /*if (fHadDecay) {
-    std::cout << " ******************************** " << std::endl
-              << " ** Initialise Pythia for e+e-->hadrons " << std::endl
-              << " ******************************** " << std::endl
-              << " Mass of the A: " << fPythia->particleData.m0(fDP) << " GeV"
-    << std::endl; fPythiaHadDecay->settings.mode("Beams:idA",  11);
-    fPythiaHadDecay->settings.mode("Beams:idB",  -11);
-    fPythiaHadDecay->settings.mode("Beams:frameType",  1);
-    fPythiaHadDecay->settings.parm("Beams:eCM",10);  // codespell:ignore parm
-    fPythiaHadDecay->readString("WeakSingleBoson:ffbar2ffbar(s:gm) = on");
-    //fPythiaHadDecay->readString("23:isResonance = true")
-     //force to hadrons
-    fPythiaHadDecay->readString("23:onMode = off");
-    fPythiaHadDecay->readString("23:onIfAny = 1 2 3 4 5");
-    }*/
+
   TDatabasePDG* pdgBase = TDatabasePDG::Instance();
   Double_t root_ctau = pdgBase->GetParticle(fDP)->Lifetime();
-  // fPythia->particleData.readString("4900023:useBreitWigner = false");
+
   if (debug) {
     std::cout << "Final particle parameters for PDGID " << fDP << ":"
               << std::endl;
@@ -147,8 +100,6 @@ Bool_t DPPythia8Generator::Init() {
   }
 
   return kTRUE;
-  // if (fHadDecay) fPythiaHadDecay->init();
-  // return kTRUE;
 }
 // -------------------------------------------------------------------------
 
@@ -158,13 +109,12 @@ DPPythia8Generator::~DPPythia8Generator() {}
 
 // -----   Passing the event   ---------------------------------------------
 Bool_t DPPythia8Generator::ReadEvent(FairPrimaryGenerator* cpg) {
+  auto [dx, dy] = CalculateBeamOffset(fsmearBeam, fPaintBeam);
   Double_t tp, tS, zp, xp, yp, zS, xS, yS, pz, px, py, e, w;
   Double_t tm, zm, xm, ym, pmz, pmx, pmy, em;
-  // Double_t td,zd,xd,yd;
+
   Int_t im;
-  // take DP decay of Pythia, move it to the SHiP decay region
-  // recalculate decay time
-  // weight event with exp(-t_ship/tau_DP) / exp(-t_pythia/tau_DP)
+
 
   int iDP =
       0;  // index of the chosen DP, also ensures that at least 1 DP is produced
@@ -172,24 +122,7 @@ Bool_t DPPythia8Generator::ReadEvent(FairPrimaryGenerator* cpg) {
       dec_chain;  // pythia indices of the particles to be stored on the stack
   std::vector<int> dpvec;  // pythia indices of DP particles
   do {
-    if (fextFile) {
-      // take charm or beauty hadron from external file
-      // correct for too much Ds produced by pythia6
-      /*bool x = true;
-        while(x){
-        if (fn==fNevents) {fLogger->Warning(MESSAGE_ORIGIN, "End of input file.
-        Rewind.");} fTree->GetEntry(fn%fNevents); fn++; if ( int(fabs(hid[0]) )
-        != 431){ x = false; } else { Double_t rnr = gRandom->Uniform(0,1); if(
-        rnr<fFDs ) { x = false; };
-        //std::cout<<"what is x "<<x<<" id "<<int(fabs(hid[0]))<<" rnr " << rnr
-        <<" "<< fFDs <<std::endl ;
-        }
-        }
-        fPythia->event.reset();
-        fPythia->event.append( (Int_t)hid[0], 1, 0, 0, hpx[0],  hpy[0],  hpz[0],
-        hE[0],  hM[0], 0., 9. );
-      */
-    }
+
     // bit for proton brem
     if (fpbrem) {
       fPythia->event.reset();
@@ -231,7 +164,7 @@ Bool_t DPPythia8Generator::ReadEvent(FairPrimaryGenerator* cpg) {
       // for mesons, could have more than one ... but for DY prod, need to take
       // the last one... int r =  int( gRandom->Uniform(0,iDP) );
       int r = iDP - 1;
-      // std::cout << " ----> debug 2 " << r  <<  std::endl;
+
       int i = dpvec[r];
       // production vertex
       zp = fPythia->event[i].zProd();
@@ -244,18 +177,14 @@ Bool_t DPPythia8Generator::ReadEvent(FairPrimaryGenerator* cpg) {
       py = fPythia->event[i].py();
       e = fPythia->event[i].e();
       // old decay vertex
-      // Int_t ida =fPythia->event[i].daughter1();
-      std::cout << " Debug: decay product of A: "
-                << fPythia->event[fPythia->event[i].daughter1()].id() << " "
-                << fPythia->event[fPythia->event[i].daughter2()].id()
-                << std::endl;
-      // hadDecay = abs(fPythia->event[ida].id())!=11 &&
-      // abs(fPythia->event[ida].id())!=13 && abs(fPythia->event[ida].id())!=15;
+      if (debug)
+	{
 
-      // zd =fPythia->event[ida].zProd();
-      // xd =fPythia->event[ida].xProd();
-      // yd =fPythia->event[ida].yProd();
-      // td =fPythia->event[ida].tProd();
+	  std::cout << " Debug: decay product of A: "
+		    << fPythia->event[fPythia->event[i].daughter1()].id() << " "
+		    << fPythia->event[fPythia->event[i].daughter2()].id()
+		    << std::endl;
+	}
       //  new decay vertex
       Double_t LS = gRandom->Uniform(fLmin, fLmax);  // mm, G4 and Pythia8 units
       Double_t p = TMath::Sqrt(px * px + py * py + pz * pz);
@@ -280,34 +209,14 @@ Bool_t DPPythia8Generator::ReadEvent(FairPrimaryGenerator* cpg) {
       pmy = fPythia->event[im].py();
       em = fPythia->event[im].e();
       tm = fPythia->event[im].tProd();
-      // foresee finite beam size
-      Double_t dx = 0;
-      Double_t dy = 0;
-      if (fsmearBeam > 0) {
-        Double_t test = fsmearBeam * fsmearBeam;
-        Double_t Rsq = test + 1.;
-        while (Rsq > test) {
-          dx = gRandom->Uniform(-1., 1.) * fsmearBeam;
-          dy = gRandom->Uniform(-1., 1.) * fsmearBeam;
-          Rsq = dx * dx + dy * dy;
-        }
-      }
-      if (fextFile) {
-        // take grand mother particle from input file, to know if primary or
-        // secondary production
-        // cpg->AddTrack((Int_t)mid[0],mpx[0],mpy[0],mpz[0],xm/cm+dx,ym/cm+dy,zm/cm,-1,false,mE[0],0.,1.);
-        // cpg->AddTrack((Int_t)fPythia->event[im].id(),pmx,pmy,pmz,xm/cm+dx,ym/cm+dy,zm/cm,0,false,em,tm/cm/c_light,w);
-        // // convert pythia's (x,y,z[mm], t[mm/c]) to ([cm], [s])
-        // cpg->AddTrack(fDP, px, py, pz, xp/cm+dx,yp/cm+dy,zp/cm,
-        // 1,false,e,tp/cm/c_light,w);
-      } else {
-        cpg->AddTrack(
-            (Int_t)fPythia->event[im].id(), pmx, pmy, pmz, xm / cm + dx,
-            ym / cm + dy, zm / cm, -1, false, em, tm / cm / c_light,
-            w);  // convert pythia's (x,y,z[mm], t[mm/c]) to ([cm], [s])
-        cpg->AddTrack(fDP, px, py, pz, xp / cm + dx, yp / cm + dy, zp / cm, 0,
-                      false, e, tp / cm / c_light, w);
-      }
+
+      cpg->AddTrack(
+		    (Int_t)fPythia->event[im].id(), pmx, pmy, pmz, (xm + dx) / cm,
+		    (ym + dy) / cm, zm / cm, -1, false, em, tm / cm / c_light,
+		    w);  // convert pythia's (x,y,z[mm], t[mm/c]) to ([cm], [s])
+      cpg->AddTrack(fDP, px, py, pz, (xp + dx) / cm , (yp + dy) / cm, zp / cm, 0,
+		    false, e, tp / cm / c_light, w);
+
       // bookkeep the indices of stored particles
       dec_chain.push_back(im);
       dec_chain.push_back(i);
@@ -355,28 +264,7 @@ Bool_t DPPythia8Generator::ReadEvent(FairPrimaryGenerator* cpg) {
     if (debug > 1) std::cout << "accept" << std::endl;
     dec_chain.push_back(k);
   }
-  //}
-  /*else {
-  //get decay from e+e- collision.....
-  fPythiaHadDecay->settings.parm("Beams:eCM",20);  // codespell:ignore parm
-  fPythiaHadDecay->next();
-  for (int k=0; k<fPythiaHadDecay->event.size(); k++){
-    fPythia->event.append(
-  fPythiaHadDecay->event[k].id(),fPythiaHadDecay->event[k].status()
-  ,fPythiaHadDecay->event[k].mother1() , fPythiaHadDecay->event[k].mother2(),
-  fPythiaHadDecay->event[k].daughter1(), fPythiaHadDecay->event[k].daughter2(),
-  fPythiaHadDecay->event[k].col(), fPythiaHadDecay->event[k].acol(),
-  fPythiaHadDecay->event[k].px(),  fPythiaHadDecay->event[k].py(),
-  fPythiaHadDecay->event[k].pz(), fPythiaHadDecay->event[k].e(),
-  fPythiaHadDecay->event[k].m(), 0., 9. ); dec_chain.push_back(
-  fPythia->event.size()-1 ); std::cout << " Adding decay product: " << k << " "
-              << fPythiaHadDecay->event[k].id() << " "
-              << fPythiaHadDecay->event[k].status() << " "
-              << fPythiaHadDecay->event[k].mother1() << " "
-              << fPythiaHadDecay->event[k].mother2() << " "
-              << std::endl;
-  }
-  }*/
+
 
   // go over daughters and store them on the stack, starting from 2 to account
   // for DP and its mother
@@ -405,9 +293,7 @@ Bool_t DPPythia8Generator::ReadEvent(FairPrimaryGenerator* cpg) {
     };
     cpg->AddTrack((Int_t)fPythia->event[k].id(), px, py, pz, xS / cm, yS / cm,
                   zS / cm, im, wanttracking, e, tS / cm / c_light, w);
-    // std::cout <<k<< " insert pdg =" <<fPythia->event[k].id() << " pz = " <<
-    // pz << " [GeV] zS = " << zS << " [mm] tS = " << tS << "[mm/c]" <<
-    // std::endl;
+
   }
   return kTRUE;
 }